+++ repos/base-linux/src/lib/base/native_thread.cc
@@ -203,7 +203,7 @@ Native_capability Native_thread::Epoll::alloc_rpc_cap()
 
 void Native_thread::Epoll::free_rpc_cap(Native_capability cap)
 {
-	int const local_socket = Capability_space::ipc_cap_data(cap).rpc_obj_key.value();
+	int const local_socket = (int)Capability_space::ipc_cap_data(cap).rpc_obj_key.value();
 
 	_exec_control([&] () { _remove(Lx_sd{local_socket}); });
 }
